extract api and update gradle build
