+++ api/build.gradle
@@ -1,0 +1,11 @@
+sourceCompatibility = "11"
+targetCompatibility = "11"
+version = "1.0-SNAPSHOT"
+
+dependencies {
+  compile project(':tictactoe')
+  testCompile group: 'junit', name: 'junit', version: '4.12'
+  compile group: 'com.fasterxml.jackson.core', name: 'jackson-core', version: '2.9.9'
+  compile group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.9.9'
+  compile group: 'com.fasterxml.jackson.core', name: 'jackson-annotations', version: '2.9.9'
+}
