+++ api/src/main/java/Encode.java
@@ -1,4 +1,3 @@
-package com.github.martingaston.tictactoe.api;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.MapperFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
