+++ api/src/test/java/APITest.java
@@ -1,5 +1,3 @@
-package com.github.martingaston.tictactoe.api;
-
 import org.junit.Test;
 
 import java.io.IOException;
