+++ api/src/test/java/EncodeTest.java
@@ -1,5 +1,3 @@
-package com.github.martingaston.tictactoe.api;
-
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.github.martingaston.tictactoe.board.Symbol;
 import org.junit.Test;
