+++ settings.gradle
@@ -11,3 +11,4 @@ rootProject.name = 'tic-tac-toe-fullstack'
 include 'tictactoe'
 include 'cli'
 include 'backend'
+include 'api'
